test: add fixtures for mock ROI points and metadata in test_structureset.py
